💥 Adapt offline pipeline for commensal operations.
This commit is a mess, since it introduces a lot of important breaking
changes at once. These include, in no particular order:

1. Remove all environment variables except `VER0_DIR`, renamed to `VER0DIR`.
2. Work according to new directory structure.
3. Enable processing multiple scans per observation.
4. Enable processing multiple observations.
5. Add command to _verify_ a run of the pipeline.
6. Add command to _kill_ a run of the pipeline.
7. Check if correlator is running, and abort if it is.
8. Inform the correlator is the pipeline is running via a flag.
9. Fix a bug wherein work was not being distributed correctly.
10. Change where pipeline logs and temporary files are written.

and many more. These changes will be stabilised over the next few
commits, and then merged into the `main` branch. More changes are on the
way as well.